change static method file
