+++ data/crac-creation/crac-creator-csa-profiles/src/test/java/com/farao_community/farao/data/crac_creation/creator/csa_profile/crac_creator/CsaProfileCracFilteringTest.java
@@ -1,0 +1,4 @@
+package com.farao_community.farao.data.crac_creation.creator.csa_profile.crac_creator;
+
+public class CsaProfileCracFilteringTest {
+}
